@@ -66,75 +74,90 @@ endif()
 find_package(cppzmq)
 
 # add include dirs
-include_directories(${PROJECT_SOURCE_DIR}/include ${OpenCV_INCLUDE_DIRS} ${NLOPT_DIR})
-if(PGR_USB3)
-    include_directories(${PGR_DIR}/include)
-    include_directories(${PGR_DIR}/include/spinnaker)
+include_directories(${PROJECT_SOURCE_DIR}/include ${OpenCV_INCLUDE_DIRS} ${NLopt_INCLUDE_DIRS})
+if(PGR_USB2 OR PGR_USB3)
+    if(MSVC)
+        include_directories(${PGR_DIR}/include)
+    else()
+        if(PGR_USB2)
+            include_directories(${PGR_DIR}/include/flycapture)   # for ubuntu default install dir
+        elseif(PGR_USB3)
+            include_directories(${PGR_DIR}/include/spinnaker)   # for ubuntu default install dir
+        endif()
+    endif()
 endif()
 
 # find sources to build
 file(GLOB LIBFICTRAC_SRCS ${PROJECT_SOURCE_DIR}/src/*.cpp)
 
 # add targets
-add_library(libfictrac STATIC ${LIBFICTRAC_SRCS})
+add_library(fictrac_core STATIC ${LIBFICTRAC_SRCS})
 add_executable(configGui ${PROJECT_SOURCE_DIR}/exec/configGui.cpp)
 add_executable(fictrac ${PROJECT_SOURCE_DIR}/exec/fictrac.cpp)
 
 # add preprocessor definitions
 # public means defs will be inherited by linked executables
-target_compile_definitions(libfictrac PUBLIC _CRT_SECURE_NO_WARNINGS NOMINMAX)
-if(PGR_USB3)
-    target_compile_definitions(libfictrac PUBLIC PGR_USB3)
+target_compile_definitions(fictrac_core PUBLIC _CRT_SECURE_NO_WARNINGS NOMINMAX)
+if(PGR_USB2)
+    target_compile_definitions(fictrac_core PUBLIC PGR_USB2)
+elseif(PGR_USB3)
+    target_compile_definitions(fictrac_core PUBLIC PGR_USB3)
 endif()
 
 # add compile options
 if(MSVC)
-    target_compile_options(libfictrac PUBLIC $<$<CONFIG:Release>:/MP /GS /GL /W3 /WX- /Gy /Zc:wchar_t /O2 /Oi /Zc:inline /fp:precise /MD /EHsc>)
+    target_compile_options(fictrac_core PUBLIC $<$<CONFIG:Release>:/MP /GS /GL /W3 /WX- /Gy /Zc:wchar_t /O2 /Oi /Zc:inline /fp:precise /MD /EHsc /std:c++17>)
+    # set(CMAKE_SHARED_LINKER_FLAGS_RELEASE "${CMAKE_SHARED_LINKER_FLAGS_RELEASE} /LTCG")
 else()  # gcc
-    target_compile_options(libfictrac PUBLIC -Ofast -Wall -c -fmessage-length=0 -std=c++14 -Wno-unused-function -march=native -MMD)
+    target_compile_options(fictrac_core PUBLIC -Ofast -Wall -c -fmessage-length=0 -std=c++17 -Wno-unused-function -march=native -MMD)
 endif()
 
-# linking and post-build
-target_link_libraries(libfictrac PUBLIC ${OpenCV_LIBS} ${NLOPT_LIB} cppzmq)
-if(PGR_USB3)
-    target_link_libraries(libfictrac PUBLIC ${PGR_LIB})
-endif()
+# linking
+target_link_libraries(fictrac_core PUBLIC ${OpenCV_LIBS} ${NLOPT_LIBRARIES} cppzmq)
 if(MSVC)
-    # win-specific libraries
-    target_link_libraries(libfictrac PUBLIC Ws2_32)
-    
+    target_link_libraries(fictrac_core PUBLIC Ws2_32)
+else()  # gcc
+    target_link_libraries(fictrac_core PUBLIC pthread)
+endif()
+if(PGR_USB2 OR PGR_USB3)
+    target_link_libraries(fictrac_core PUBLIC ${PGR_LIB})
+endif()
+
+# post-build
+if(MSVC)
     # copy all opencv dlls
     set(OPENCV_VER_STRING ${OpenCV_VERSION_MAJOR}${OpenCV_VERSION_MINOR}${OpenCV_VERSION_PATCH})
     foreach(lib ${OpenCV_LIBS})
-        if(EXISTS "${_OpenCV_LIB_PATH}/${lib}${OPENCV_VER_STRING}.dll")
-            list(APPEND TO_COPY "${_OpenCV_LIB_PATH}/${lib}${OPENCV_VER_STRING}.dll" "${_OpenCV_LIB_PATH}/${lib}${OPENCV_VER_STRING}d.dll")
+        if(EXISTS "${OpenCV_INSTALL_PATH}/bin/${lib}${OPENCV_VER_STRING}.dll")
+            list(APPEND TO_COPY "${OpenCV_INSTALL_PATH}/bin/${lib}${OPENCV_VER_STRING}.dll")
         endif()
     endforeach()
     
-    set(FFMPEG_LIB_BASE opencv_ffmpeg${OPENCV_VER_STRING})
-    if(${OpenCV_ARCH} STREQUAL x64)
-        set(FFMPEG_LIB ${FFMPEG_LIB_BASE}_64.dll)
-    else()
-        set(FFMPEG_LIB ${FFMPEG_LIB_BASE}.dll)
-    endif()
-    list(APPEND TO_COPY "${_OpenCV_LIB_PATH}/${FFMPEG_LIB}")
+    # set(FFMPEG_LIB_BASE opencv_ffmpeg${OPENCV_VER_STRING})
+    # if("${OpenCV_ARCH}" STREQUAL x86)
+    #     set(FFMPEG_LIB ${FFMPEG_LIB_BASE}.dll)
+    # else()
+    #     # default to 64-bit
+    #     set(FFMPEG_LIB ${FFMPEG_LIB_BASE}_64.dll)
+    # endif()
+    # list(APPEND TO_COPY "${OpenCV_INSTALL_PATH}/bin/${FFMPEG_LIB}")
     
     # copy nlopt dll
-    list(APPEND TO_COPY "${NLOPT_DIR}/libnlopt-0.dll")
+    # list(APPEND TO_COPY "${NLOPT_LIBRARY_DIRS}/bin/nlopt.dll")
     
     # copy h264 dll
-    list(APPEND TO_COPY "${PROJECT_SOURCE_DIR}/dll/openh264-1.7.0-win64.dll")
+    file(GLOB DLLS ${PROJECT_SOURCE_DIR}/dll/*.dll)
+    list(APPEND TO_COPY "${DLLS}")
     
     add_custom_command( TARGET fictrac POST_BUILD
                         COMMAND ${CMAKE_COMMAND} -E copy_if_different
                         ${TO_COPY}
                         "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/$<$<CONFIG:Release>:Release>$<$<CONFIG:Debug>:Debug>/")
 else()  # gcc
-    # socket libs, pthread
-    target_link_libraries(libfictrac PUBLIC pthread)
+    # nothing here...
 endif()
 
-target_link_libraries(configGui libfictrac)
-add_dependencies(configGui libfictrac)
-target_link_libraries(fictrac libfictrac)
-add_dependencies(fictrac libfictrac)
+target_link_libraries(configGui fictrac_core)
+add_dependencies(configGui fictrac_core)
+target_link_libraries(fictrac fictrac_core)
+add_dependencies(fictrac fictrac_core)
