@@ -18,8 +18,8 @@ Logger::Logger()
 {
     // create log writer
     string fn = string("fictrac-") + execTime() + ".log";
-    _log = unique_ptr<Recorder>(new Recorder(RecorderInterface::RecordType::FILE, fn));
-    _cout = unique_ptr<Recorder>(new Recorder(RecorderInterface::RecordType::TERM));
+    _log = make_unique<Recorder>(RecorderInterface::RecordType::FILE, fn);
+    _cout = make_unique<Recorder>(RecorderInterface::RecordType::TERM);
     if (_log->is_active() && _cout->is_active()) {
         cout << "Initialised logging to " << fn << endl;
     } else {
@@ -61,22 +61,21 @@ void Logger::mprintf(LogLevel lvl, string func, string format, ...)
     // not re-entrant
     lock_guard<mutex> l1(log._pMutex);
 
+    // expand args
+    va_list args;
+    va_start(args, format);
+    vsnprintf(buf, buf_size, format.c_str(), args);
+    va_end(args);
+
     // print and log
     if ((int)lvl >= (int)verbosity()) {
-        
-        // expand args
-        va_list args;
-        va_start(args, format);
-        vsnprintf(buf, buf_size, format.c_str(), args);
-        va_end(args);
-
         // async printing to console
         log._cout->addMsg(string(buf) + "\n");
+    }
 
-        // don't log display text to file
-        if (lvl != PRT) {
-            // async logging to file (with additional info)
-            log._log->addMsg(to_string(elapsed_secs()) + " " + func + " [" + log.LogLevelStrings[lvl] + "] " + buf + "\n");
-        }
+    // don't log display text to file (but log everything else)
+    if (lvl != PRT) {
+        // async logging to file (with additional info)
+        log._log->addMsg(to_string(elapsed_secs()) + " " + func + " [" + log.LogLevelStrings[lvl] + "] " + buf + "\n");
     }
 }
