@@ -13,6 +13,7 @@
 #include <opencv2/opencv.hpp>
 
 #include <memory>   // shared_ptr
+#include <vector>
 
 ///
 ///
@@ -22,7 +23,7 @@ class Localiser : public NLoptFunc
 public:
     Localiser(nlopt_algorithm alg, double bound, double tol, int max_evals,
         CameraModelPtr sphere_model, const cv::Mat& sphere_map,
-        const cv::Mat& roi_mask, std::shared_ptr<double[]> p1s_lut);
+        const cv::Mat& roi_mask, std::shared_ptr<std::vector<double>> p1s_lut);
     ~Localiser() {};
 
     double search(cv::Mat& roi_frame, cv::Mat& R_roi, CmPoint64f& vx);
@@ -36,7 +37,7 @@ private:
     const double* _R_roi;
     CameraModelPtr _sphere_model;
     const cv::Mat _sphere_map, _roi_mask;
-    std::shared_ptr<double[]> _p1s_lut;
+    std::shared_ptr<std::vector<double>> _p1s_lut;
     cv::Mat _roi_frame;
     int _roi_w, _roi_h;
 };
