@@ -24,12 +24,14 @@ using cv::Mat;
 /// Constructor.
 ///
 CVSource::CVSource(std::string input)
+    : _is_image(false)
 {
     LOG_DBG("Source is: %s", input.c_str());
     Mat test_frame;
     try {
         // try reading input as camera id
-        LOG_DBG("Trying source as camera id..");
+        LOG_DBG("Trying source as camera id...");
+        if (input.size() > 2) { throw std::exception(); }
         int id = std::stoi(input);
         _cap = std::shared_ptr<cv::VideoCapture>(new cv::VideoCapture(id));
         if (!_cap->isOpened()) { throw 0; }
@@ -42,7 +44,7 @@ CVSource::CVSource(std::string input)
     catch (...) {
         try {
             // then try loading as video file
-            LOG_DBG("Trying source as video file..");
+            LOG_DBG("Trying source as video file...");
             _cap = std::shared_ptr<cv::VideoCapture>(new cv::VideoCapture(input));
             if (!_cap->isOpened()) { throw 0; }
             *_cap >> test_frame;
@@ -52,14 +54,42 @@ CVSource::CVSource(std::string input)
             _live = false;
         }
         catch (...) {
-            LOG_ERR("Could not interpret source type (%s)!", input.c_str());
-            _open = false;
+            try {
+                // then try loading as an image file
+                LOG_DBG("Trying source as image file...");
+                _frame_cap = cv::imread(input);
+                if (_frame_cap.empty()) { throw 0; }
+                LOG("Using source type: image file.");
+                _open = true;
+                _live = false;
+                _is_image = true;
+            }
+            catch (...) {
+                LOG_ERR("Could not interpret source type (%s)!", input.c_str());
+                _open = false;
+            }
         }
     }
 
 	if( _open ) {
-		_width = static_cast<int>(_cap->get(cv::CAP_PROP_FRAME_WIDTH));
-		_height = static_cast<int>(_cap->get(cv::CAP_PROP_FRAME_HEIGHT));
+        if (_is_image) {
+            _width = _frame_cap.cols;
+            _height = _frame_cap.rows;
+        }
+        else {
+            _width = static_cast<int>(_cap->get(cv::CAP_PROP_FRAME_WIDTH));
+            _height = static_cast<int>(_cap->get(cv::CAP_PROP_FRAME_HEIGHT));
+        }
+        if (_live) {
+            _fps = getFPS();    // don't init fps for video files - we might want to play them back as fast as possible
+
+            LOG("OpenCV camera source initialised (%dx%d @ %.3f fps)!", _width, _height, _fps);
+        }
+        else if (_is_image) {
+            LOG("OpenCV image source initialised (%dx%d)!", _width, _height);
+        } else {
+            LOG("OpenCV video source initialised (%dx%d)!", _width, _height);
+        }
 	}
 }
 
@@ -74,10 +104,11 @@ CVSource::~CVSource()
 ///
 double CVSource::getFPS()
 {
-    if (_open) {
-        _fps = _cap->get(cv::CAP_PROP_FPS);
+    double fps = _fps;
+    if (_open && _cap) {
+        fps = _cap->get(cv::CAP_PROP_FPS);
     }
-    return _fps;
+    return fps;
 }
 
 ///
@@ -86,12 +117,16 @@ double CVSource::getFPS()
 bool CVSource::setFPS(double fps)
 {
     bool ret = false;
-    if (_open && (fps > 0)) {
-        _fps = fps;
-        if (!_cap->set(cv::CAP_PROP_FPS, _fps)) {
-            LOG_WRN("Warning! Failed to set device fps (attempted to set fps=%.2f).", _fps);
+    if (_open && _cap && (fps > 0)) {
+        if (!_cap->set(cv::CAP_PROP_FPS, fps)) {
+            LOG_WRN("Warning! Failed to set device fps (attempted to set fps=%.2f).", fps);
+            _fps = fps; // just set fps anyway for playback
+            LOG("Playback frame rate is now %.2f", _fps);
         }
-        else { ret = true; }
+        else {
+            _fps = getFPS();
+            LOG("Device frame rate is now %.2f", _fps);
+        }
     }
     return ret;
 }
@@ -103,7 +138,7 @@ bool CVSource::setFPS(double fps)
 bool CVSource::rewind()
 {
     bool ret = false;
-	if (_open) {
+	if (_open && _cap) {
         if (!_cap->set(cv::CAP_PROP_POS_FRAMES, 0)) {
             LOG_WRN("Warning! Failed to rewind source.");
         } else { ret = true; }
@@ -117,12 +152,14 @@ bool CVSource::rewind()
 bool CVSource::grab(cv::Mat& frame)
 {
 	if( !_open ) { return false; }
-	if( !_cap->read(_frame_cap) ) {
+	if( !_is_image && !_cap->read(_frame_cap) ) {
 		LOG_ERR("Error grabbing image frame!");
 		return false;
 	}
-    double ts = static_cast<double>(ts_ms());    // backup, in case the device timestamp is junk
+    double ts = ts_ms();    // backup, in case the device timestamp is junk
+    _ms_since_midnight = ms_since_midnight();
 	_timestamp = _cap->get(cv::CAP_PROP_POS_MSEC);
+    LOG_DBG("Frame captured %dx%d%d @ %f (t_sys: %f ms, t_day: %f ms)", _frame_cap.cols, _frame_cap.rows, _frame_cap.channels(), _timestamp, ts, _ms_since_midnight);
     if (_timestamp <= 0) {
         _timestamp = ts;
     }
@@ -130,20 +167,20 @@ bool CVSource::grab(cv::Mat& frame)
 	if( _frame_cap.channels() == 1 ) {
 		switch( _bayerType ) {
 			case BAYER_BGGR:
-				cv::cvtColor(_frame_cap, frame, CV_BayerBG2BGR);
+				cv::cvtColor(_frame_cap, frame, cv::COLOR_BayerBG2BGR);
 				break;
 			case BAYER_GBRG:
-				cv::cvtColor(_frame_cap, frame, CV_BayerGB2BGR);
+				cv::cvtColor(_frame_cap, frame, cv::COLOR_BayerGB2BGR);
 				break;
 			case BAYER_GRBG:
-				cv::cvtColor(_frame_cap, frame, CV_BayerGR2BGR);
+				cv::cvtColor(_frame_cap, frame, cv::COLOR_BayerGR2BGR);
 				break;
 			case BAYER_RGGB:
-				cv::cvtColor(_frame_cap, frame, CV_BayerRG2BGR);
+				cv::cvtColor(_frame_cap, frame, cv::COLOR_BayerRG2BGR);
 				break;
 			case BAYER_NONE:
 			default:
-				cv::cvtColor(_frame_cap, frame, CV_GRAY2BGR);
+				cv::cvtColor(_frame_cap, frame, cv::COLOR_GRAY2BGR);
 				break;
 		}
 	} else {
@@ -152,9 +189,9 @@ bool CVSource::grab(cv::Mat& frame)
 
     /// Correct average frame rate when reading from file.
     if (!_live && (_fps > 0)) {
-        static double prev_ts = ts - 25; // initially 40 Hz
-        static double av_fps = 40;      // initially 40 Hz
-        static double sleep_ms = 25;
+        static double prev_ts = ts - (1000/_fps);
+        static double av_fps = _fps;      // initially 40 Hz
+        static double sleep_ms = 1000/_fps;
         av_fps = 0.15 * av_fps + 0.85 * (1000 / (ts - prev_ts));
         sleep_ms *= 0.25 * (av_fps / _fps) + 0.75;
         sleep(static_cast<long>(round(sleep_ms)));
