@@ -18,9 +18,9 @@
 #include "timing.h"
 #include "misc.h"
 #include "CVSource.h"
-#ifdef PGR_USB3
+#if defined(PGR_USB2) || defined(PGR_USB3)
 #include "PGRSource.h"
-#endif // PGR_USB3
+#endif // PGR_USB2/3
 
 /// OpenCV individual includes required by gcc?
 #include <opencv2/highgui.hpp>
@@ -141,7 +141,7 @@ void createZoomROI(Mat& zoom_roi, const Mat& frame, const Point2d& pt, int orig_
     int x = frame.cols/2;
     if (pt.x >= 0) { x = clamp(int(pt.x - orig_dim/2 + 0.5), int(orig_dim/2), frame.cols - 1 - orig_dim); }
     int y = frame.rows/2;
-    if (pt.y >= 0) { y = clamp(int(pt.y - orig_dim/2 + 0.5), int(orig_dim/2), frame.rows - 1 - orig_dim); }
+    if (pt.y >= 0) { y = clamp(int(pt.y - orig_dim/2 + 0.5), 0, frame.rows - 1 - orig_dim); }
     Mat crop_rect = frame(cv::Rect(x, y, orig_dim, orig_dim));
     cv::resize(crop_rect, zoom_roi, zoom_roi.size());
 }
@@ -168,8 +168,9 @@ ConfigGui::ConfigGui(string config_fn)
     Mat input_frame;
     std::shared_ptr<FrameSource> source;
     if (_open) {
-#ifdef PGR_USB3
+#if defined(PGR_USB2) || defined(PGR_USB3)
         try {
+            if (input_fn.size() > 2) { throw std::exception(); }
             // first try reading input as camera id
             int id = std::stoi(input_fn);
             source = std::make_shared<PGRSource>(id);
@@ -178,9 +179,9 @@ ConfigGui::ConfigGui(string config_fn)
             // then try loading as video file
             source = std::make_shared<CVSource>(input_fn);
         }
-#else // PGR_USB3
+#else // !PGR_USB2/3
         source = std::make_shared<CVSource>(input_fn);
-#endif // PGR_USB3
+#endif // PGR_USB2/3
         if (!source->isOpen()) {
             LOG_ERR("Error! Could not open input frame source (%s)!", input_fn.c_str());
             _open = false;
@@ -190,6 +191,28 @@ ConfigGui::ConfigGui(string config_fn)
         } else if (input_frame.empty()) {
             _open = false;
         }
+    }
+
+    /// Optionally enhance frame for config
+    bool do_enhance = false;
+    _cfg.getBool("enh_cfg_disp", do_enhance);
+    if (_open && do_enhance) {
+        LOG("Enhancing config image ..");
+        Mat maximg = input_frame.clone();
+        Mat minimg = input_frame.clone();
+        while (source->grab(input_frame)) {
+            for (int i = 0; i < input_frame.rows; i++) {
+                uint8_t* pmin = minimg.ptr(i);
+                uint8_t* pmax = maximg.ptr(i);
+                const uint8_t* pimg = input_frame.ptr(i);
+                for (int j = 0; j < input_frame.cols * input_frame.channels(); j++) {
+                    uint8_t p = pimg[j];
+                    if (p > pmax[j]) { pmax[j] = p; }
+                    if (p < pmin[j]) { pmin[j] = p; }
+                }
+            }
+        }
+        input_frame = maximg - minimg;
     }
 
     /// Create base file name for output files.
@@ -225,7 +248,7 @@ bool ConfigGui::setFrame(Mat& frame)
         _frame = frame.clone();
     } else if (frame.channels() == 1) {
         //_frame = frame.clone();
-        cv::cvtColor(frame, _frame, CV_GRAY2BGR);
+        cv::cvtColor(frame, _frame, cv::COLOR_GRAY2BGR);
     } else {
         // uh oh, shouldn't get here
         LOG_ERR("Unexpected number of image channels (%d)!", frame.channels());
@@ -247,34 +270,25 @@ bool ConfigGui::setFrame(Mat& frame)
 	}
 
     LOG("Using vfov: %f deg", vfov);
-    
-    //FIXME: support also fisheye models!
-    _cam_model = CameraModel::createRectilinear(static_cast<int>(_w), static_cast<int>(_h), vfov * CM_D2R);
+
+    bool fisheye = false;
+    if (_cfg.getBool("fisheye", fisheye) && fisheye) {
+        _cam_model = CameraModel::createFisheye(_w, _h, vfov * CM_D2R / (double)_h, 360 * CM_D2R);
+    }
+    else {
+        // default to rectilinear
+        _cam_model = CameraModel::createRectilinear(_w, _h, vfov * CM_D2R);
+    }
     
     return true;
 }
 
 ///
 /// Write camera-animal transform to config file.
+/// Warning: input R+t is animal to camera frame transform!
 ///
-bool ConfigGui::saveC2ATransform(const Mat& R, const Mat& t)
+bool ConfigGui::saveC2ATransform(const string& ref_str, const Mat& R, const Mat& t)
 {
-	string sqr_type = "";
-	switch (_input_data.mode) {
-		case R_XY:
-			sqr_type = "c2a_cnrs_xy";
-			break;
-		case R_YZ:
-			sqr_type = "c2a_cnrs_yz";
-			break;
-		case R_XZ:
-			sqr_type = "c2a_cnrs_xz";
-			break;
-        default:
-            LOG_ERR("Uh oh, something went wrong :-(");
-            return false;
-	}
-
 	// dump corner points to config file
 	vector<int> cfg_pts;
 	for (auto p : _input_data.sqrPts) {
@@ -283,13 +297,13 @@ bool ConfigGui::saveC2ATransform(const Mat& R, const Mat& t)
 	}
 
 	// write to config file
-	LOG("Adding c2a_src and %s to config file and writing to disk (%s) ..", sqr_type.c_str(), _config_fn.c_str());
-    _cfg.add("c2a_src", sqr_type);
-    _cfg.add(sqr_type, cfg_pts);
+	LOG("Adding c2a_src and %s to config file and writing to disk (%s) ..", ref_str.c_str(), _config_fn.c_str());
+    _cfg.add("c2a_src", ref_str);
+    _cfg.add(ref_str, cfg_pts);
 
 	// dump R to config file
 	vector<double> cfg_r, cfg_t;
-	CmPoint angleAxis = CmPoint64f::matrixToOmega(R);
+	CmPoint angleAxis = CmPoint64f::matrixToOmega(R.t());   // transpose to get camera-animal transform
 	for (int i = 0; i < 3; i++) {
 		cfg_r.push_back(angleAxis[i]);
 		cfg_t.push_back(t.at<double>(i, 0));
@@ -320,21 +334,25 @@ bool ConfigGui::saveC2ATransform(const Mat& R, const Mat& t)
 ///
 /// Update animal coordinate frame estimate.
 ///
-bool ConfigGui::updateC2ATransform(const Mat& ref_cnrs, Mat& R, Mat& t)
+bool ConfigGui::updateRt(const string& ref_str, Mat& R, Mat& t)
 {
-	bool ret = false;
-	if (_input_data.newEvent) {
-        //FIXME: also support edge clicks! e.g.:
-        //  double x1 = click[2 * i + 0].x;     double y1 = click[2 * i + 0].y;
-        //  double x2 = click[2 * i + 1].x;     double y2 = click[2 * i + 1].y;
-        //  double x3 = click[2 * i + 2].x;     double y3 = click[2 * i + 2].y;
-        //  double x4 = click[2 * i + 3].x;     double y4 = click[2 * i + 3].y;
-        //  double px = ((x1*y2 - y1 * x2)*(x3 - x4) - (x1 - x2)*(x3*y4 - y3 * x4)) / ((x1 - x2)*(y3 - y4) - (y1 - y2)*(x3 - x4));
-        //  double py = ((x1*y2 - y1 * x2)*(y3 - y4) - (y1 - y2)*(x3*y4 - y3 * x4)) / ((x1 - x2)*(y3 - y4) - (y1 - y2)*(x3 - x4));
-		ret = computeRtFromSquare(_cam_model, ref_cnrs, _input_data.sqrPts, R, t);
-		_input_data.newEvent = false;
-	}
-	return ret;
+    //FIXME: also support edge clicks! e.g.:
+    //  double x1 = click[2 * i + 0].x;     double y1 = click[2 * i + 0].y;
+    //  double x2 = click[2 * i + 1].x;     double y2 = click[2 * i + 1].y;
+    //  double x3 = click[2 * i + 2].x;     double y3 = click[2 * i + 2].y;
+    //  double x4 = click[2 * i + 3].x;     double y4 = click[2 * i + 3].y;
+    //  double px = ((x1*y2 - y1 * x2)*(x3 - x4) - (x1 - x2)*(x3*y4 - y3 * x4)) / ((x1 - x2)*(y3 - y4) - (y1 - y2)*(x3 - x4));
+    //  double py = ((x1*y2 - y1 * x2)*(y3 - y4) - (y1 - y2)*(x3*y4 - y3 * x4)) / ((x1 - x2)*(y3 - y4) - (y1 - y2)*(x3 - x4));
+
+    bool ret = false;
+    if (ref_str == "c2a_cnrs_xy") {
+        ret = computeRtFromSquare(_cam_model, XY_CNRS, _input_data.sqrPts, R, t);
+    } else if (ref_str == "c2a_cnrs_yz") {
+        ret = computeRtFromSquare(_cam_model, YZ_CNRS, _input_data.sqrPts, R, t);
+    } else if (ref_str == "c2a_cnrs_xz") {
+        ret = computeRtFromSquare(_cam_model, XZ_CNRS, _input_data.sqrPts, R, t);
+    }
+    return ret;
 }
 
 /////
@@ -363,11 +381,22 @@ bool ConfigGui::updateC2ATransform(const Mat& ref_cnrs, Mat& R, Mat& t)
 ///
 ///
 ///
-void ConfigGui::drawC2ACorners(Mat& disp_frame, const Mat& ref_cnrs, const Mat& R, const Mat& t)
+void ConfigGui::drawC2ACorners(Mat& disp_frame, const string& ref_str, const Mat& R, const Mat& t)
 {
     // make x4 mat for projecting corners
     Mat T(3, 4, CV_64F);
     for (int i = 0; i < 4; i++) { t.copyTo(T.col(i)); }
+
+    Mat ref_cnrs;
+    if (ref_str == "c2a_cnrs_xy") {
+        ref_cnrs = XY_CNRS;
+    } else if (ref_str == "c2a_cnrs_yz") {
+        ref_cnrs = YZ_CNRS;
+    } else if (ref_str == "c2a_cnrs_xz") {
+        ref_cnrs = XZ_CNRS;
+    } else {
+        return;
+    }
 
     // project reference corners
     Mat p = R * ref_cnrs + T;
@@ -386,6 +415,7 @@ void ConfigGui::drawC2AAxes(Mat& disp_frame, const Mat& R, const Mat& t, const d
         double scale = 1.0 / tan(r);
         Mat so = (cv::Mat_<double>(3, 1) << c.x, c.y, c.z) * scale;
         drawAxes(disp_frame, _cam_model, R, so, Scalar(0, 0, 255));
+        drawAnimalAxis(disp_frame, _cam_model, R, so, r, Scalar(255, 0, 0));
     }
 }
 
@@ -407,6 +437,10 @@ bool ConfigGui::run()
     /// Interactive window.
     cv::namedWindow("configGUI", cv::WINDOW_AUTOSIZE);
     cv::setMouseCallback("configGUI", onMouseEvent, &_input_data);
+
+    /// If reconfiguring, then delete pre-computed values.
+    bool reconfig = false;
+    _cfg.getBool("reconfig", reconfig);
     
     /// Display/input loop.
 	Mat R, t;
@@ -414,24 +448,25 @@ bool ConfigGui::run()
     double r = -1;
     char key = 0;
     string val;
-	string cfg_r_src;
+	string c2a_src;
     vector<int> cfg_pts;
     vector<double> cfg_vec;
     vector<vector<int>> cfg_polys;
 	changeState(CIRC_INIT);
-    const char exit_key = 0x1b;
-#ifdef WIN32
-    const char enter_key = 0x0d;
-#else // WIN32
-    const char enter_key = 0x0a;
-#endif // WIN32
     const int click_rad = std::max(int(_w/150+0.5), 5);
     Mat disp_frame, zoom_frame(ZOOM_DIM, ZOOM_DIM, CV_8UC3);
     const int scaled_zoom_dim = static_cast<int>(ZOOM_DIM * ZOOM_SCL + 0.5);
-    while (_open && (key != exit_key)) {
+    while (_open && (key != 0x1b)) {    // esc
         /// Create frame for drawing.
         //cv::cvtColor(_frame, disp_frame, CV_GRAY2RGB);
         disp_frame = _frame.clone();
+
+        // normalise zoom window
+        {
+            double min, max;
+            cv::minMaxLoc(disp_frame, &min, &max);
+            disp_frame = (disp_frame - min) * 255 / (max - min);
+        }
         
         int in;
         string str;
@@ -442,10 +477,10 @@ bool ConfigGui::run()
             
                 // test read
                 cfg_pts.clear();
-                if (_cfg.getVecDbl("roi_c", cfg_vec) && _cfg.getDbl("roi_r", r)) {
+                if (!reconfig && _cfg.getVecDbl("roi_c", cfg_vec) && _cfg.getDbl("roi_r", r)) {
                     c.copy(cfg_vec.data());
                     LOG_DBG("Found roi_c = [%f %f %f] and roi_r = %f rad.", c[0], c[1], c[2], r);
-                    LOG_WRN("Warning! When roi_c and roi_r are specified in the config file, roi_circ will be ignored.\nTo re-compute roi_c and roi_r, please delete these values from the config file and reconfigure.");
+                    LOG_WRN("Warning! When roi_c and roi_r are specified in the config file, roi_circ will be ignored.\nTo re-compute roi_c and roi_r, please delete these values or set reconfig : y in the config file and reconfigure.");
                 } 
                 else if (_cfg.getVecInt("roi_circ", cfg_pts)) {
 
@@ -457,9 +492,25 @@ bool ConfigGui::run()
 
                     /// Fit circular FoV to sphere.
                     if (_input_data.circPts.size() >= 3) {
-                        circleFit_camModel(_input_data.circPts, _cam_model, c, r);
+                        if (circleFit_camModel(_input_data.circPts, _cam_model, c, r)) {
 
-                        LOG_DBG("Computed roi_c = [%f %f %f] and roi_r = %f rad from %d roi_circ points.", c[0], c[1], c[2], r, _input_data.circPts.size());
+                            LOG_DBG("Computed roi_c = [%f %f %f] and roi_r = %f rad from %d roi_circ points.", c[0], c[1], c[2], r, _input_data.circPts.size());
+
+                            // save re-computed values
+                            cfg_vec.clear();
+                            cfg_vec.push_back(c[0]);
+                            cfg_vec.push_back(c[1]);
+                            cfg_vec.push_back(c[2]);
+
+                            // write to config file
+                            LOG("Adding roi_c and roi_r to config file and writing to disk (%s) ..", _config_fn.c_str());
+                            _cfg.add("roi_c", cfg_vec);
+                            _cfg.add("roi_r", r);
+                            if (_cfg.write() <= 0) {
+                                LOG_ERR("Error writing to config file (%s)!", _config_fn.c_str());
+                                _open = false;  // will cause exit
+                            }
+                        }
                     }
                 }
                 else {
@@ -470,7 +521,7 @@ bool ConfigGui::run()
                 /// Draw fitted circumference.
                 if (r > 0) {
                     drawCircle_camModel(disp_frame, _cam_model, c, r, Scalar(255,0,0), false);
-                            
+        
                     /// Display.
                     cv::imshow("configGUI", disp_frame);
                     cv::waitKey(100);   //FIXME: why do we have to wait so long to make sure the frame is drawn?
@@ -480,7 +531,7 @@ bool ConfigGui::run()
                     // input loop
                     while (true) {
                         cv::waitKey(100);   //FIXME: dirty hack - sometimes image doesn't draw, at least with this line we can just mash keys until it does
-                        printf("\n  Would you like to keep the existing spherer ROI configuration ([y]/n)? ");
+                        printf("\n  Would you like to keep the existing sphere ROI configuration ([y]/n)? ");
                         in = getchar();
                         switch (in)
                         {
@@ -527,7 +578,7 @@ bool ConfigGui::run()
                 
                 /// Draw previous clicks.
                 for (auto click : _input_data.circPts) {
-                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, CV_AA);
+                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, cv::LINE_AA);
                 }
                 
                 /// Draw fitted circumference.
@@ -545,7 +596,7 @@ bool ConfigGui::run()
                 key = cv::waitKey(5);
                 
                 /// State machine logic.
-                if (key == enter_key) {
+                if ((key == 0x0d) || (key == 0x0a)) {   // return
                     if (_input_data.circPts.size() >= 3) {
                         // dump circumference points, c, and r to config file
                         cfg_pts.clear();
@@ -604,9 +655,9 @@ bool ConfigGui::run()
                     for (unsigned int i = 0; i < _input_data.ignrPts.size(); i++) {
                         for (unsigned int j = 0; j < _input_data.ignrPts[i].size(); j++) {
                             if (i == _input_data.ignrPts.size()-1) {
-                                cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, CV_AA);
+                                cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
                             }
-                            cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j+1)%_input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, CV_AA);
+                            cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j+1)%_input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
                         }
                     }
                     
@@ -657,9 +708,9 @@ bool ConfigGui::run()
                 for (unsigned int i = 0; i < _input_data.ignrPts.size(); i++) {
                     for (unsigned int j = 0; j < _input_data.ignrPts[i].size(); j++) {
                         if (i == _input_data.ignrPts.size()-1) {
-                            cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, CV_AA);
+                            cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
                         }
-                        cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j+1)%_input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, CV_AA);
+                        cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j+1)%_input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
                     }
                 }
                 
@@ -678,7 +729,7 @@ bool ConfigGui::run()
                 key = cv::waitKey(5);
                 
                 /// State machine logic.
-                if (key == enter_key) {
+                if ((key == 0x0d) || (key == 0x0a)) {  // return
                     // if current poly is empty, assume we've finished
                     if (_input_data.ignrPts.empty() || _input_data.ignrPts.back().empty()) {
                         if (!_input_data.ignrPts.empty()) { _input_data.ignrPts.pop_back(); }
@@ -720,60 +771,69 @@ bool ConfigGui::run()
             
             /// Choose method for defining animal frame.
             case R_INIT:
-                /// Load R+t transform from config file.
-                cfg_vec.clear();
-                if (_cfg.getVecDbl("c2a_r", cfg_vec)) {
-                    LOG_DBG("Read c2a_r = [%f %f %f]", cfg_vec[0], cfg_vec[1], cfg_vec[2]);
-                    R = CmPoint64f::omegaToMatrix(CmPoint(cfg_vec[0], cfg_vec[1], cfg_vec[2]));
-                }
-                else {
-                    LOG_DBG("Error reading c2a_r from config file! Re-running configuration ..");
-                    changeState(R_SLCT);
-                    break;
+                /// Check if corners specified (optional).
+                _input_data.sqrPts.clear();
+                if (_cfg.getStr("c2a_src", c2a_src)) {
+                    LOG_DBG("Found c2a_src: %s", c2a_src.c_str());
+
+                    /// Load square corners from config file.
+                    cfg_pts.clear();
+                    if (_cfg.getVecInt(c2a_src, cfg_pts)) {
+                        for (unsigned int i = 1; i < cfg_pts.size(); i += 2) {
+                            _input_data.sqrPts.push_back(cv::Point2d(cfg_pts[i - 1], cfg_pts[i]));
+                        }
+                    }
                 }
 
+                /// Load R+t transform from config file.
+                R.release();    // clear mat
                 cfg_vec.clear();
-                if (_cfg.getVecDbl("c2a_t", cfg_vec)) {
+                if (!reconfig && _cfg.getVecDbl("c2a_r", cfg_vec)) {
+                    LOG_DBG("Read c2a_r = [%f %f %f]", cfg_vec[0], cfg_vec[1], cfg_vec[2]);
+                    R = CmPoint64f::omegaToMatrix(CmPoint(cfg_vec[0], cfg_vec[1], cfg_vec[2])).t();     // transpose to lab-camera transform
+                }
+                else {
+                    LOG_WRN("Warning! c2a_r missing from config file. Looking for corner points..");
+                }
+
+                t.release();    // clear mat
+                cfg_vec.clear();
+                if (!reconfig && _cfg.getVecDbl("c2a_t", cfg_vec)) {
                     LOG_DBG("Read c2a_t = [%f %f %f]", cfg_vec[0], cfg_vec[1], cfg_vec[2]);
                     t = (cv::Mat_<double>(3, 1) << cfg_vec[0], cfg_vec[1], cfg_vec[2]);
                 }
                 else {
-                    LOG_DBG("Error reading c2a_t from config file! Re-running configuration ..");
+                    LOG_WRN("Warning! c2a_t missing from config file. Looking for corner points..");
+                }
+
+                if (R.empty() || t.empty()) {
+                    if (!_input_data.sqrPts.empty()) {
+                        LOG_DBG("Recomputing R+t from specified corner points...");
+
+                        /// Recompute R+t
+                        if (updateRt(c2a_src, R, t)) {
+                            saveC2ATransform(c2a_src, R, t);
+                        }
+                    }
+                }
+                else {
+                    LOG_WRN("Warning! When c2a_r and c2a_t are specified in the config file, c2a_src and associated corners points will be ignored.\nTo re-compute c2a_r and c2a_t, please delete these values or set reconfig : y in the config file and reconfigure.");
+                }
+
+                /// If c2a_r/t missing and couldn't re-compute from specified corners points.
+                if (R.empty() || t.empty()) {
+                    LOG_ERR("Error! Could not read or compute c2a_r and/or c2a_t. Re-running configuration..");
                     changeState(R_SLCT);
                     break;
                 }
 
-                LOG_WRN("Warning! When c2a_r and c2a_t are specified in the config file, c2a_src and associated corners points will be ignored.\nTo re-compute c2a_r and c2a_t, please delete these values from the config file and reconfigure.");
-
-                /// Check also if corners specified (unnecessary).
-                if (_cfg.getStr("c2a_src", cfg_r_src)) {
-                    LOG_DBG("Found c2a_src: %s", cfg_r_src.c_str());
-
-                    /// Load square corners from config file.
-                    cfg_pts.clear();
-                    if (_cfg.getVecInt(cfg_r_src, cfg_pts)) {
-                        _input_data.sqrPts.clear();
-                        for (unsigned int i = 1; i < cfg_pts.size(); i += 2) {
-                            _input_data.sqrPts.push_back(cv::Point2d(cfg_pts[i - 1], cfg_pts[i]));
-                        }
-
-                        /// Draw previous clicks.
-                        for (auto click : _input_data.sqrPts) {
-                            cv::circle(disp_frame, click, click_rad, Scalar(255, 255, 0), 1, CV_AA);
-                        }
-
-                        /// Draw reference corners.
-                        if (cfg_r_src == "c2a_cnrs_xy") {
-                            drawC2ACorners(disp_frame, XY_CNRS, R, t);
-                        }
-                        else if (cfg_r_src == "c2a_cnrs_yz") {
-                            drawC2ACorners(disp_frame, YZ_CNRS, R, t);
-                        }
-                        else if (cfg_r_src == "c2a_cnrs_xz") {
-                            drawC2ACorners(disp_frame, XZ_CNRS, R, t);
-                        }
-                    }
+                /// Draw previous clicks.
+                for (auto click : _input_data.sqrPts) {
+                    cv::circle(disp_frame, click, click_rad, Scalar(255, 255, 0), 1, cv::LINE_AA);
                 }
+
+                /// Draw reference corners.
+                drawC2ACorners(disp_frame, c2a_src, R, t);
 
                 /// Draw axes.
                 drawC2AAxes(disp_frame, R, t, r, c);
@@ -846,18 +906,21 @@ bool ConfigGui::run()
                     {
                         case 1:
                             printf("\n\n\n  XY-square method.\n\n  Please click on the four corners of a square shape that is aligned with the animal's X-Y axes. The corners must be clicked in the following order: (+X,-Y), (+X,+Y), (-X,+Y), (-X,-Y). If your camera is looking down on the animal from above, then the four corners are (in order): TL, TR, BR, BL from the camera's perspective. If your camera is below the animal, then the order is TR, TL, BL, BR.\n\n  Make sure the displayed axis is the correct right-handed coordinate frame!!\n\n  You can hold F to mirror the axis if the handedness is incorrect.\n\n  Press ENTER when you are satisfied with the animal's axis, or press ESC to exit..\n\n");
+                            c2a_src = "c2a_cnrs_xy";
                             // advance state
 							changeState(R_XY);
                             break;
                             
                         case 2:
                             printf("\n\n\n  YZ-square method.\n\n  Please click on the four corners of a square shape that is aligned with the animal's Y-Z axes. The corners must be clicked in the following order: (-Y,-Z), (+Y,-Z), (+Y,+Z), (-Y,+Z). If your camera is behind the animal, then the four corners are (in order): TL, TR, BR, BL from the camera's perspective. If your camera is in front of the animal, then the order is TR, TL, BL, BR.\n\n  Make sure the displayed axis is the correct right-handed coordinate frame!!\n\n  You can hold F to mirror the axis if the handedness is incorrect.\n\n  Press ENTER when you are satisfied with the animal's axis, or press ESC to exit..\n\n");
+                            c2a_src = "c2a_cnrs_yz";
                             // advance state
 							changeState(R_YZ);
                             break;
                             
                         case 3:
                             printf("\n\n\n  XZ-square method.\n\n  Please click on the four corners of a square shape that is aligned with the animal's X-Z axes. The corners must be clicked in the following order: (+X,-Z), (-X,-Z), (-X,+Z), (+X,+Z). If your camera is to the animal's left side, then the four corners are (in order): TL, TR, BR, BL from the camera's perspective. If your camera is to the animal's right side, then the order is TR, TL, BL, BR.\n\n  Make sure the displayed axis is the correct right-handed coordinate frame!!\n\n  You can hold F to mirror the axis if the handedness is incorrect.\n\n  Press ENTER when you are satisfied with the animal's axis, or press ESC to exit..\n\n");
+                            c2a_src = "c2a_cnrs_xz";
                             // advance state
 							changeState(R_XZ);
                             break;
@@ -869,6 +932,7 @@ bool ConfigGui::run()
                             // break;
                             
                         case 5:
+                            c2a_src = "ext";
                             // advance state
 							changeState(R_EXT);
                             break;
@@ -887,13 +951,16 @@ bool ConfigGui::run()
             
                 /// Draw previous clicks.
                 for (auto click : _input_data.sqrPts) {
-                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, CV_AA);
+                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, cv::LINE_AA);
                 }
                 
                 /// Draw axes.
                 if (_input_data.sqrPts.size() == 4) {
-					updateC2ATransform(XY_CNRS, R, t);
-                    drawC2ACorners(disp_frame, XY_CNRS, R, t);
+                    if (_input_data.newEvent) {
+                        updateRt(c2a_src, R, t);
+                        _input_data.newEvent = false;
+                    }
+                    drawC2ACorners(disp_frame, c2a_src, R, t);
                     drawC2AAxes(disp_frame, R, t, r, c);
                 }
                 
@@ -909,10 +976,10 @@ bool ConfigGui::run()
                 key = cv::waitKey(5);
                 
                 /// State machine logic.
-                if (key == enter_key) {
+                if ((key == 0x0d) || (key == 0x0a)) {   // return
                     if ((_input_data.sqrPts.size() == 4) && !R.empty()) {
                         // dump corner points to config file
-						if (!saveC2ATransform(R, t)) {
+						if (!saveC2ATransform(c2a_src, R, t)) {
 							LOG_ERR("Error writing coordinate transform to config file!");
                             _open = false;      // will cause exit
 						}
@@ -937,13 +1004,16 @@ bool ConfigGui::run()
                 
                 /// Draw previous clicks.
                 for (auto click : _input_data.sqrPts) {
-                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, CV_AA);
+                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, cv::LINE_AA);
                 }
                 
                 /// Draw axes.
                 if (_input_data.sqrPts.size() == 4) {
-					updateC2ATransform(YZ_CNRS, R, t);
-                    drawC2ACorners(disp_frame, YZ_CNRS, R, t);
+                    if (_input_data.newEvent) {
+                        updateRt(c2a_src, R, t);
+                        _input_data.newEvent = false;
+                    }
+                    drawC2ACorners(disp_frame, c2a_src, R, t);
                     drawC2AAxes(disp_frame, R, t, r, c);
                 }
                 
@@ -959,10 +1029,10 @@ bool ConfigGui::run()
                 key = cv::waitKey(5);
                 
                 /// State machine logic.
-                if (key == enter_key) {
+                if ((key == 0x0d) || (key == 0x0a)) {   // return
 					if ((_input_data.sqrPts.size() == 4) && !R.empty()) {
                         // dump corner points to config file
-						if (!saveC2ATransform(R, t)) {
+						if (!saveC2ATransform(c2a_src, R, t)) {
 							LOG_ERR("Error writing coordinate transform to config file!");
                             _open = false;      // will cause exit
 						}
@@ -987,13 +1057,16 @@ bool ConfigGui::run()
                 
                 /// Draw previous clicks.
                 for (auto click : _input_data.sqrPts) {
-                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, CV_AA);
+                    cv::circle(disp_frame, click, click_rad, Scalar(255,255,0), 1, cv::LINE_AA);
                 }
                 
                 /// Draw axes.
                 if (_input_data.sqrPts.size() == 4) {
-					updateC2ATransform(XZ_CNRS, R, t);
-                    drawC2ACorners(disp_frame, XZ_CNRS, R, t);
+                    if (_input_data.newEvent) {
+                        updateRt(c2a_src, R, t);
+                        _input_data.newEvent = false;
+                    }
+                    drawC2ACorners(disp_frame, c2a_src, R, t);
                     drawC2AAxes(disp_frame, R, t, r, c);
                 }
                 
@@ -1009,10 +1082,10 @@ bool ConfigGui::run()
                 key = cv::waitKey(5);
                 
                 /// State machine logic.
-                if (key == enter_key) {
+                if ((key == 0x0d) || (key == 0x0a)) {   // return
 					if ((_input_data.sqrPts.size() == 4) && !R.empty()) {
                         // dump corner points to config file
-						if (!saveC2ATransform(R, t)) {
+						if (!saveC2ATransform(c2a_src, R, t)) {
 							LOG_ERR("Error writing coordinate transform to config file!");
                             _open = false;      // will cause exit
 						}
@@ -1080,7 +1153,7 @@ bool ConfigGui::run()
             
             /// Exit config.
             case EXIT:
-                key = exit_key;
+                key = 0x1b; // esc
                 break;
         }
     }
@@ -1100,21 +1173,15 @@ bool ConfigGui::run()
 	for (unsigned int i = 0; i < _input_data.ignrPts.size(); i++) {
 		for (unsigned int j = 0; j < _input_data.ignrPts[i].size(); j++) {
 			if (i == _input_data.ignrPts.size() - 1) {
-				cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, CV_AA);
+				cv::circle(disp_frame, _input_data.ignrPts[i][j], click_rad, COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
 			}
-			cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j + 1) % _input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, CV_AA);
+			cv::line(disp_frame, _input_data.ignrPts[i][j], _input_data.ignrPts[i][(j + 1) % _input_data.ignrPts[i].size()], COLOURS[i%NCOLOURS], 1, cv::LINE_AA);
 		}
 	}
 
 	// draw animal coordinate frame
 	if (_input_data.sqrPts.size() == 4) {
-		if (cfg_r_src == "c2a_cnrs_xy") {
-            drawC2ACorners(disp_frame, XY_CNRS, R, t);
-		} else if (cfg_r_src == "c2a_cnrs_yz") {
-            drawC2ACorners(disp_frame, YZ_CNRS, R, t);
-		} else if (cfg_r_src == "c2a_cnrs_xz") {
-            drawC2ACorners(disp_frame, XZ_CNRS, R, t);
-		}
+        drawC2ACorners(disp_frame, c2a_src, R, t);
 	}
     drawC2AAxes(disp_frame, R, t, r, c);
 
