@@ -11,7 +11,6 @@
 #include <cstdio>
 #include <iostream>
 #include <fstream>
-#include <sstream>
 #include <exception>    // try, catch
 #include <algorithm>    // erase, remove
 
@@ -71,20 +70,20 @@ int ConfigParser::read(string fn)
         }
 
         /// Tokenise
-        const string whitespace = ", \t\n\r";
+        const string whitespace = ", \t\n";
         std::size_t delim = line.find(":");
         if (delim >= line.size()) { continue; } // skip blank lines
         string key = line.substr(0, line.find_last_not_of(whitespace, delim - 1) + 1), val = "";
         try {
             val = line.substr(line.find_first_not_of(whitespace, delim + 1));
-            //val.erase(std::remove(val.begin(), val.end(), '\r'), val.end());    // remove /r under linux
+            val.erase(std::remove(val.begin(), val.end(), '\r'), val.end());    // remove /r under linux
         }
         catch (...) {}  // add blank values
 
         /// Add to map
         _data[key] = val;
 
-        LOG_DBG("Extracted key: %s  val: %s", key.c_str(), val.c_str());
+        LOG_DBG("Extracted key: |%s|  val: |%s|", key.c_str(), val.c_str());
     }
 
     /// Clean up
@@ -124,9 +123,11 @@ int ConfigParser::write(string fn)
     }
 
     /// Write comments
-    f << std::endl;
-    for (auto c : _comments) {
-        f << c << std::endl;
+    if (_comments.size() > 0) {
+        f << std::endl;
+        for (auto c : _comments) {
+            f << c << std::endl;
+        }
     }
 
     /// Clean up
@@ -141,29 +142,15 @@ int ConfigParser::write(string fn)
 ///
 ///
 ///
-string ConfigParser::operator()(string key)
+string ConfigParser::operator()(string key) const
 {
-    string s = "";
-    getStr(key, s);
-    return s;
-}
-
-///
-///
-///
-template<typename T>
-T ConfigParser::get(string key)
-{
-    T val;
-    string s;
-    if (getStr(key, s)) {
-        std::stringstream ss(s);
-        try { ss >> val; }
-        catch (std::exception& e) {
-            LOG_ERR("Error parsing config file value (%s : %s)! Error was: %s", key.c_str(), ss.str().c_str(), e.what());
-        }
+    try {
+        return _data.at(key);
     }
-    return val;
+    catch (...) {
+        LOG_DBG("Key (%s) not found.", key.c_str());
+    }
+    return "";
 }
 
 ///
@@ -175,7 +162,7 @@ bool ConfigParser::getStr(string key, string& val) {
         val = _data[key];
         return true;
     }
-    LOG_WRN("Warning! Key (%s) not found.", key.c_str());
+    LOG_DBG("Key (%s) not found.", key.c_str());
     return false;
 }
 
